requirements.txt
Django==2.2.2
djangorestframework==3.10.0
nltk==3.4.4
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
